+++ proxy/user.js
@@ -3,3 +3,20 @@ var user = require('../model/user');
 exports.getUsers = function (callback) {
     user.find({}).limit(20).sort({ follower: -1}).exec(callback);
 };
+
+exports.getUserById = function (id, callback) {
+    user.find({id: id}, callback);
+};
+
+exports.saveUser = function (article, following, follower, callback) {
+    user.create({
+        id: article.authorHref,
+        author: article.author,
+        following: following,
+        follower: follower
+    }, callback)
+};
+
+exports.updateUser = function (article, following, follower, callback) {
+    user.update({id: article.authorHref}, {following: following, follower: follower},callback)
+};