+++ routes/user.js
@@ -5,6 +5,7 @@ var user = require('../proxy/user');
 /* GET user listing. */
 router.get('/', function(req, res, next) {
     user.getUsers(function (err, users) {
+        if (err) return next(err);
         res.render('user', {users: users});
     });
 });